setup cleanup the code
