--- API_projTraining/API_projTraining.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31321.278
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "API_projTraining", "API_projTraining.csproj", "{318BC8F5-B9DC-4739-A7FB-2A579053B5D8}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{318BC8F5-B9DC-4739-A7FB-2A579053B5D8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{318BC8F5-B9DC-4739-A7FB-2A579053B5D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{318BC8F5-B9DC-4739-A7FB-2A579053B5D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{318BC8F5-B9DC-4739-A7FB-2A579053B5D8}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E73C3CFD-AA98-4636-AD83-34787AE2D848}
-	EndGlobalSection
-EndGlobal
