--- API_projTraining/Libraries/User.cs
@@ -1,17 +1,0 @@
-﻿namespace API_projTraining.Services
-{
-    public class User
-    {
-        public int Id { get; set; }
-
-        public string Email { get; set; }
-
-        public string Password { get; set; }
-
-        public long Telephone { get; set; }
-
-        public string FirstName { get; set; }
-
-        public string LastName { get; set; }
-    }
-}
