+++ src/API_projTraining/Controllers/UsersController.cs
@@ -1,25 +1,68 @@
 ﻿using API_projTraining.Services;
 using Microsoft.AspNetCore.Mvc;
-using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
 namespace API_projTraining.Controllers
 {
-    [Route("[Users]")]
     [ApiController]
-    public class UsersController : ControllerBase
+    [Route("api/User")]
+    //[Consumes("application/json")]
+    public class ActionResult : ControllerBase
     {
-        // GET: <UsersController>
-        [HttpGet]
-        public IEnumerable<UsersController> Get()
+        [HttpGet("OneUser")]
+        public ActionResult<User> FindUser(int id)
         {
-            Users.Equals
-                
+            UserServices myuserservices = new();
+            if (id > 0 & id < 11)
+            {
+                return myuserservices.FindUser(id);
+            }
+            else
+            {
+                return NotFound();
+            }
         }
 
+        [HttpGet("AllUsers")]
+        public List<User> FindAllUsers()
+        {
+            UserServices myuserservices = new();
+            return myuserservices.FindAllUsers();
+        }
+
+        [HttpGet("SomeUsers")]
+        //need to sort a way to fetch and return random users from the array
+        public ActionResult<List<User>> FindSomeUsers(int numberOfUsers)
+        {
+            UserServices myuserservices = new();
+            if (numberOfUsers > 0 & numberOfUsers < 11)
+            {
+                return myuserservices.FindSomeUsers(numberOfUsers);
+            }
+            else
+            {
+                return NotFound();
+            }
+        }
+
+        //// POST api/<UsersController>
+        //[HttpPost]
+        //public void Post([FromBody] string value)
+        //{
+        //}
+
+        //// PUT api/<UsersController>/5
+        //[HttpPut("{id}")]
+        //public void Put(int id, [FromBody] string value)
+        //{
+        //}
+
+        //// DELETE api/<UsersController>/5
+        //[HttpDelete("{id}")]
+        //public void Delete(int id)
+        //{
+        //}
     }
 }
