+++ src/API_projTraining/Library/Messages.cs
@@ -1,0 +1,10 @@
+﻿namespace API_projTraining.Services
+{
+    public class Messages
+    {
+        public int UserId { get; set; }
+        public int MsgId { get; set; }
+        public string Subject { get; set; }
+        public string Body { get; set; }
+    }
+}
