+++ src/API_projTraining/Library/User.cs
@@ -1,0 +1,17 @@
+﻿namespace API_projTraining.Services
+{
+    public class User
+    {
+        public int Id { get; set; }
+
+        public string Email { get; set; }
+
+        public string Password { get; set; }
+
+        public long Telephone { get; set; }
+
+        public string FirstName { get; set; }
+
+        public string LastName { get; set; }
+    }
+}
